requirements.txt
Django
psycopg2-binary
aiohttp==3.1.2
lxml
cssselect
py-dateutil
scikit-learn
numpy
scipy
nltk